+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Add tabbed navigation components to the UI library
     * Create the `Tab` component
-    * Export the `Tabs` MUI component
+    * Create the `Tabs` component
 
 # 2.29.0 (2021-02-23)